First refactor, TRIOMACLASS
